requirements.txt
fastapi==0.115.0
uvicorn==0.41.0
sqlalchemy==2.0.36
databases==0.9.0
faster-whisper==1.2.1
sumy==0.12.0
nltk==3.9.3
pydantic-settings==2.13.1
python-multipart==0.0.22
pytest==8.3.4
httpx==0.28.1
requests==2.32.5
python-dotenv==1.0.0
numpy
av
